python_extractor/requirements.txt
google-genai
PyPDF2
langchain-text-splitters
